# Conflicts:
#	app/src/androidTest/java/com/android/shelfLife/ui/EndToEndM1Test.kt
#	app/src/androidTest/java/di/TestUserRepositoryModule.kt
#	app/src/androidTest/java/helpers/FoodItemRepositoryTestHelper.kt
#	app/src/androidTest/java/helpers/HouseholdRepositoryTestHelper.kt
#	app/src/androidTest/java/helpers/UserRepositoryTestHelper.kt
#	app/src/main/java/com/android/shelfLife/viewmodel/camera/BarcodeScannerViewModel.kt

+++ app/src/main/java/com/android/shelfLife/viewmodel/overview/IndividualFoodItemViewModel.kt
@@ -25,7 +25,7 @@ constructor(
 
   suspend fun deleteFoodItem() {
     val householdId = userRepository.user.value?.selectedHouseholdUID
-    if (householdId != null) {
+    if (householdId != null && selectedFood != null) {
       foodItemRepository.deleteFoodItem(householdId, selectedFood!!.uid)
       foodItemRepository.selectFoodItem(null)
     }
